Remove unused interfaces
