--- SubtitleFileCleanerGUI/Model/IDeformatable.cs
@@ -1,8 +1,0 @@
-﻿namespace SubtitleFileCleanerGUI.Model
-{
-    // Interface describes an object that stores information about the need to remove tags
-    public interface IDeformatable
-    {
-        public bool DeleteTags { get; set; }
-    }
-}
