--- SubtitleFileCleanerGUI/Model/IDislocatable.cs
@@ -1,8 +1,0 @@
-﻿namespace SubtitleFileCleanerGUI.Model
-{
-    // Interface describes an object that stores information about the final destination path
-    public interface IDislocatable
-    {
-        public string PathDestination { get; set; }
-    }
-}
