--- SubtitleFileCleanerGUI/Model/ILocatable.cs
@@ -1,8 +1,0 @@
-﻿namespace SubtitleFileCleanerGUI.Model
-{
-    // Interface describes an object that stores the location path.
-    public interface ILocatable
-    {
-        public string PathLocation { get; set; }
-    }
-}
